Initial Commit
This is the first cut of the project, and I have still to make sure it
does all taht I want it to do, but that's coming right up. For now, we
can easily handle small registrations, or passing in functions to pull
up the list of URLs for posting to, and all kinds of options. This is
pretty close to what I wanted.