Added the docs for the project
I really needed to add in the docs for how to use this library and
examples of how you might implement the more useful features on the
global registrations. This should be ready to go - all I need now are
the tests of things to make sure it's all really working.