Finished up the Clojars work
I wanted to add in the `flush!` function, and that turns out to be
easiest if added as a simple `delete!` on the queues and then it's just
all done. I also updated the README with the directions for the
inclusion in a project, and we should be ready to go to Clojars.